% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCheckDuplicates.R
\name{codedAsInteger}
\alias{codedAsInteger}
\title{codedAsInteger}
\usage{
codedAsInteger(input)
}
\arguments{
\item{input}{Input data that should be checked, provided as a object related to a data frame.}
}
\value{
values encoded as integers without NAs.
}
\description{
Encodes data as integer values(corresponding to their factor representatives).
NA values are exchanged by 0.
}
