% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectFunctions.R
\name{collectFunctions}
\alias{collectFunctions}
\title{collectFunctions}
\usage{
collectFunctions(pattern = "^check", globalenv = FALSE,
  allowed_args = c("x", "mx", "cfg"))
}
\arguments{
\item{pattern}{Name pattern the function name should match}

\item{globalenv}{Boolean deciding whether functions in the global environment should be considered
or not.}

\item{allowed_args}{Vector of allowed arguments. If a function contains an argument not listed here it will be
ignored and a warning will be returned.}
}
\value{
a character vector of function calls fulfilling all requirements
}
\description{
Collects functions which follow a given name pattern
and compares their arguments against a list of allowed arguments
}
\examples{

collectFunctions()

}
\seealso{
\code{\link{iamCheck}}
}
\author{
Jan Philipp Dietrich
}
