% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputData.R
\name{createInputData}
\alias{createInputData}
\title{createInputData}
\usage{
createInputData(x, cfg = "CDLINKS", ref = "IAMC", verbose = TRUE, ...)
}
\arguments{
\item{x}{Input data that should be checked, provided as a file path to a reporting file,
a quitte object or an object which can be converted to quitte using \code{\link[quitte]{as.quitte}}}

\item{cfg}{Project configuration that should be used (currently available: "CDLINKS"). Either a project name, a path to a
config file or a data frame specifying available variables and corresponding properties as returned by
\code{\link{iamProjectConfig}()}.}

\item{ref}{Reference data for comparison. Either a project name (currently available: "IAMC"), a path to a mif
file or a quitte object containing the data.}

\item{verbose}{Boolean influencing the degree of information returned by the function. \code{verbose=TRUE} returns
detailed information whereas \code{verbose=FALSE} returns a summary.}

\item{...}{additional data objects which are forwarded to the check functions}
}
\value{
Input named list with elements available for check functions
}
\description{
Takes given data and provides it in the required data format (magpie)
}
\author{
Cornelia Auer
}
