% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_landcover.R
\name{example_landcover}
\alias{example_landcover}
\title{example_landcover}
\value{
land cover data in quitte format
}
\description{
Example data from CDLINKS containing land cover information. For testing some mistakes
have been introduced which will trigger warnings in iamCheck
}
\author{
Jan Philipp Dietrich
}
