% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_magpie.R
\name{example_magpie}
\alias{example_magpie}
\title{example_magpie}
\value{
unspecified MAgPIE model outputs in quitte format
}
\description{
Example output from the MAgPIE model
}
\author{
Jan Philipp Dietrich
}
