% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterInputData.R
\name{filterInputData}
\alias{filterInputData}
\title{filterInputData}
\usage{
filterInputData(input, cfg = "CDLINKS", globalenv = FALSE, out = NULL)
}
\arguments{
\item{input}{Named list with elements available for check functions}

\item{cfg}{Project configuration that should be used. Either a project name (currently available: "CDLINKS"), a path to a
config file or a data frame specifying available variables and corresponding properties as returned by
\code{\link{iamProjectConfig}()}.}

\item{globalenv}{Boolean deciding whether functions in the global environment should be considered}

\item{out}{List with status from pre-checks, e.g. illegal or missing variables.}
}
\value{
List with 1) filtered input and 2) status-output about the consistency of the input data
}
\description{
Filters given data (input) according to available default variables (cfg).
Pre-checks are performed about the consistency of the input data (e.g. illegal or missing variables)
and the status (out) returned.
}
\author{
Cornelia Auer
}
