% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCheckDuplicates.R
\name{getShiftFactor}
\alias{getShiftFactor}
\title{getShiftFactor}
\usage{
getShiftFactor(positiveNumber)
}
\arguments{
\item{positiveNumber}{number >= 0}
}
\value{
order of power of ten for the input value
}
\description{
Method: Essentially computes order by the power of ten from a given positive number
The order value is then used to compute a shift factor, which is needed to uniquely represent the data values as integer in preCheckDuplicates.
}
