% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamCheck.R
\name{iamCheck}
\alias{iamCheck}
\title{iamCheck}
\usage{
iamCheck(
  x,
  pdf = NULL,
  cfg = "CDLINKS",
  refData = "IAMC",
  verbose = FALSE,
  globalenv = FALSE,
  pdfStyle = NULL,
  ...
)
}
\arguments{
\item{x}{Input data that should be checked, provided as a file path to a reporting file,
a quitte object or an object which can be converted to quitte using \code{\link[quitte]{as.quitte}}}

\item{pdf}{File name used for a PDF containing diagnostic results of the check. If set to NULL
no pdf will be written.}

\item{cfg}{Project configuration that should be used. Either a project name (currently available: "CDLINKS"), a path to a
config file or a data frame specifying available variables and corresponding properties as returned by
\code{\link{iamProjectConfig}()}.}

\item{refData}{Reference data for comparison. Either a project name (currently available: "IAMC"), a path to a mif
file or a quitte object containing the data.}

\item{verbose}{Boolean influencing the degree of information returned by the function. \code{verbose=TRUE} returns
detailed information whereas \code{verbose=FALSE} returns a summary.}

\item{globalenv}{Boolean deciding whether functions in the global environment should be considered
or not.}

\item{pdfStyle}{list of style-options for the pdf}

\item{...}{additional data objects which are forwarded to the check functions}
}
\value{
List of all inputs and outputs created by the performed checks (invisible)
}
\description{
Runs various diagnostics over a provided data set checking whether
the provided data is in line with IAMC database guidelines.
}
\examples{

# run check with example data
iamCheck(example_REMIND, cfg="CDLINKS")

}
\seealso{
\code{\link{iamProjectConfig}}, \code{\link[quitte]{as.quitte}}, \code{\link[quitte]{is.quitte}}
}
\author{
Jan Philipp Dietrich
}
