% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamReferenceData.R
\name{iamReferenceData}
\alias{iamReferenceData}
\title{iamReferenceData}
\usage{
iamReferenceData(ref = "IAMC")
}
\arguments{
\item{ref}{Reference data for comparison. Either a project name (currently available: "IAMC"), a path to a
file or a data frame containing the data.}
}
\value{
Quitte object containing the reference data set
}
\description{
Function to return historical reference data for validation.
}
\examples{
iamReferenceData()

}
\seealso{
\code{\link[quitte]{as.quitte}}, \code{\link[quitte]{is.quitte}}
}
\author{
Jan Philipp Dietrich
}
