% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamSummaryPDF.R
\name{iamSummaryPDF}
\alias{iamSummaryPDF}
\title{iamSummaryPDF}
\usage{
iamSummaryPDF(input, check_results = NULL, file = "summary.pdf")
}
\arguments{
\item{input}{named list with elements available for check functions}

\item{check_results}{list with check results as returned by \code{\link{iamCheck}}}

\item{file}{File name the summary should be written to.}
}
\description{
Creates a PDF summarizing check results and adding
additional validation output
}
\examples{
\dontrun{
input <- list(x=example_landcover, val=iamValidationData())
check_results <- iamCheck(example_landcover)

iamSummaryPDF(input, check_results)
}
}
\seealso{
\code{\link{iamCheck}}, \code{\link{iamProjectConfig}}
}
\author{
Jan Philipp Dietrich
}
