% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamValidationData.R
\name{iamValidationData}
\alias{iamValidationData}
\title{iamValidationData}
\usage{
iamValidationData(val = "IAMC")
}
\arguments{
\item{val}{Validation data for comparison. Either a project name (currently available: "IAMC"), a path to a
file or a data frame containing the data.}
}
\value{
quitte object containing the validation data set
}
\description{
Function to return historical data for validation.
}
\examples{
iamValidationData()

}
\seealso{
\code{\link[quitte]{as.quitte}}, \code{\link[quitte]{is.quitte}}
}
\author{
Jan Philipp Dietrich
}
