% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamVariables.R
\name{iamVariables}
\alias{iamVariables}
\title{iamVariables}
\usage{
iamVariables(cfg = "IAMC")
}
\arguments{
\item{cfg}{Configuration that should be used. Either a project name (currently available: "IAMC"), a path to a
config file or a data frame specifying available variables and corresponding properties}
}
\value{
data frame containing avaible variables and corresponding properties
}
\description{
Function to return available variables and corresponding properties for a given configuration (e.g. project).
}
\examples{
iamVariables()

}
\seealso{
\code{\link[quitte]{as.quitte}}, \code{\link[quitte]{is.quitte}}
}
\author{
Jan Philipp Dietrich
}
