% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCheckDuplicates.R
\name{preCheckDuplicates}
\alias{preCheckDuplicates}
\title{preCheckDuplicates}
\usage{
preCheckDuplicates(x)
}
\arguments{
\item{x}{Input data that should be checked, provided as a quitte object \code{\link[quitte]{as.quitte}}}
}
\value{
Two-dimensional list. 1) message with how many dpulicates 2) list of duplicates (in their integer representation)
}
\description{
Checks for duplicate entries in the data.
To optimize performance the single data values are encoded into integer values (corresponding to their factor representatives).
The encoded integer values are then compared if duplicates exist.
}
\author{
Cornelia Auer
}
