% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCheck.R
\name{processCheck}
\alias{processCheck}
\title{processCheck}
\usage{
processCheck(check, verbose = TRUE)
}
\arguments{
\item{verbose}{Boolean influencing the degree of information returned by the function. \code{verbose=TRUE} returns
detailed information whereas \code{verbose=FALSE} returns a summary.}
}
\value{
summary of check results
}
\description{
Processes a check, which means that it returns a note if the check fails, returns a check summary based on the chosen
verbosity and returns the check results in a structured way. All checks should be run within processCheck.
Requirement for a check function is that it returns a list with 2 elements: "message" which contains the standard message
that should show up for the test and "failed" which is a vector of names for which the corresponding test failed.
}
\examples{

processCheck(iamc:::checkVariable(example_magpie, cfg=iamVariables()))

}
\seealso{
\code{\link{iamVariables}}, \code{\link[quitte]{as.quitte}}, \code{\link[quitte]{is.quitte}}
}
\author{
Jan Philipp Dietrich
}
