% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterInputData.R
\name{filterInputData}
\alias{filterInputData}
\title{filterInputData}
\usage{
filterInputData(input, cfg = "CDLINKS", globalenv = FALSE, out = NULL)
}
\arguments{
\item{input}{named list with elements available for check functions}

\item{cfg}{Project configuration that should be used. Either a project name (currently available: "CDLINKS"), a path to a
config file or a data frame specifying available variables and corresponding properties as returned by
\code{\link{iamProjectConfig}()}.}

\item{globalenv}{Boolean deciding whether functions in the global environment should be considered}

\item{out}{list to be filled for status-output}
}
\value{
list with 1) filtered input and 2) status-output about illegal or missing variables
}
\description{
Filters given data (input) according to available default variables and
returns status (out) about illegal or missing variables
}
\author{
Cornelia Auer
}
