% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamCheck.R
\name{iamCheck}
\alias{iamCheck}
\title{iamCheck}
\usage{
iamCheck(x, pdf = NULL, cfg = "IAMC", val = "IAMC", verbose = TRUE)
}
\arguments{
\item{x}{Input data that should be checked, provided as a file path to a reporting file,
a quitte object or an object which can be converted to quitte using \code{\link[quitte]{as.quitte}}}

\item{pdf}{File name used for a PDF containing diagnostic results of the check. If set to NULL
no pdf will be written.}

\item{cfg}{Configuration that should be used. Either a project name (currently available: "IAMC"), a path to a
config file or a data frame specifying available variables and corresponding properties as returned by
\code{\link{iamVariables}()}.}

\item{val}{Validation data for comparison. Either a project name (currently available: "IAMC"), a path to a mif
file or a quitte object containing the data.}

\item{verbose}{Boolean influencing the degree of information returned by the function. \code{verbose=TRUE} returns
detailed information whereas \code{verbose=FALSE} returns a summary.}
}
\value{
list of all outputs created by the performed checks
}
\description{
Runs various diagnostics over a provided data set checking whether
the provided data is in line with IAMC database guidelines.
}
\examples{

# run check with example data
iamCheck(example_magpie)

}
\seealso{
\code{\link{iamVariables}}, \code{\link[quitte]{as.quitte}}, \code{\link[quitte]{is.quitte}}
}
\author{
Jan Philipp Dietrich
}
