% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamProjectConfig.R
\name{iamProjectConfig}
\alias{iamProjectConfig}
\title{iamProjectConfig}
\usage{
iamProjectConfig(cfg = "CDLINKS")
}
\arguments{
\item{cfg}{Project configuration that should be used. Either a project name (currently available: "CDLINKS"), a path to a
config file or a data frame specifying available variables and corresponding properties}
}
\value{
Data frame containing avaible variables and corresponding properties
}
\description{
Function to return available variables and corresponding properties for a given project configuration.
}
\examples{
iamProjectConfig()

}
\seealso{
\code{\link[quitte]{as.quitte}}, \code{\link[quitte]{is.quitte}}
}
\author{
Jan Philipp Dietrich
}
