% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamSummaryPDF.R
\name{iamSummaryPDF}
\alias{iamSummaryPDF}
\title{iamSummaryPDF}
\usage{
iamSummaryPDF(
  input,
  check_results = NULL,
  file = "summary.pdf",
  maxLinesOutput = 200,
  pdfStyle = NULL,
  ...
)
}
\arguments{
\item{input}{named list with elements available for check functions}

\item{check_results}{list with check results as returned by \code{\link{iamCheck}}}

\item{file}{File name the summary should be written to or a Sweave object. If a sweave object is provided the function will return the updated object, otherwise it will write its content to the file.}

\item{maxLinesOutput}{maximum number of lines that should be output in the pdf}

\item{pdfStyle}{list of style-options for the pdf}

\item{...}{additional arguments sent to \code{\link[lusweave]{swclose}}}
}
\description{
Creates a PDF summarizing check results and adding additional results for a
comparison to reference data (e.g. matching to historical data)
}
\examples{
\dontrun{
input <- list(x=example_REMIND, ref=iamReferenceData())
check_results <- iamCheck(example_REMIND)

iamSummaryPDF(check_results$input, check_results$out)
}
}
\seealso{
\code{\link{iamCheck}}, \code{\link{iamProjectConfig}}
}
\author{
Jan Philipp Dietrich
}
