% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCheck.R
\name{processCheck}
\alias{processCheck}
\title{processCheck}
\usage{
processCheck(check, env)
}
\arguments{
\item{check}{function call that should be run}

\item{env}{Environment containing all relevant settings but also outputs of previous checks.
Outputs of the given check function will be added to the environment}
}
\value{
environment containing all settings and outputs
}
\description{
Processes a check, which means that it returns a note if the check fails, returns a check summary based on the chosen
verbosity and returns the check results in a structured way. All checks should be run within processCheck.
Requirement for a check function is that it returns a list with 2 elements: "message" which contains the standard message
that should show up for the test and "failed" which is a vector of names for which the corresponding test failed.
}
\examples{

e <- processCheck(iamc:::checkVariable(example_magpie, cfg=iamVariables()), env=new.env())
e$out
rm(out, envir=e)
}
\seealso{
\code{\link{iamVariables}}, \code{\link[quitte]{as.quitte}}, \code{\link[quitte]{is.quitte}}
}
\author{
Jan Philipp Dietrich
}
